requirements.txt
django == 4.2.5
djangorestframework
django-cors-headers
django-environ
djangorestframework-simplejwt
psycopg2-binary
xhtml2pdf
openpyxl
